Move CombatTurn and Modifier classes to their own files
